faux-rm to add new characters individually and initialize db
